Feat complete purchasing process front end (#11)
ShoppingCartPage shows the current content of the cart. The employee can increase/decrease the number of items, complete the order or go back to the ProductListPage (in creating order mode).
When an order is confirmed, the employee will be redirected to the ClientDetailsPage connected to the client for who he has just completed the order.
ShoppingCartPage contains these components:

ShoppingCartTitle, displays the name of the client
ShoppingCartiTable, displays the content of the cart
ShoppingCartControls, displays two buttons, one for confirming the order and the other one for going back to the ProductListPage
ShoppingCartTotAmount, displays the current total amount of the order
ModalOrderConfirmed, displays a modal for checking if the employee is sure that he wants to send the order.
It has been added also the CreateNewOrderButton which appears in the ClientDetailsPage and if pressed it redirects to the ProductListPage in creating order mode.

The flow of purchasing process is: from ClientDetails an employee clicks on "Create new order" button, he is redirected to ProductListPage in creating order mode, where he can click on "Add to cart" on each item and add a specific quantity to that item in the ShoppingCart. When he wants he can go to the ShoppingCartPage where he can consult the current content of the cart and send the order.

It refers to the task [S202109SPG-16](https://polito-se2-21-09.myjetbrains.com/youtrack/issue/S202109SPG-16)