Purchase Process for a client - front end (#48)
A client must be logged in order to create an order. So if not logged he sees ProductListPage in non-creating order mode (without add-to-cart button), the same page that an employee sees at this path "/". 
The process of filling the shopping cart is the same followed by the employee: when the client thinks to be ready to send the order he goes to the ShoppingCartPage by clicking the cart icon on the NavbarComponent and from here he sends the request for completeOrder API. 
Then, he/she is redirected to his/her page  ("/client") where a confirm for order successfully sent appears. 

Removed logic from components.

Task on you track: [Implement client order - front end](https://polito-se2-21-09.myjetbrains.com/youtrack/issue/S202109SPG-34)